requirements.txt
streamlit==1.32.0
faster-whisper==0.10.0
numpy==1.26.4
soundfile==0.12.1
python-docx==1.1.0 